+++ src/config/import-build-target.ts
@@ -1,10 +1,8 @@
 import { isAdmin } from './index';
 
-const importBuildTarget = () => {
+export const importBuildTarget = () => {
   if (isAdmin) {
     return import('../components/app/app-admin');
   }
   return import('../components/app/app');
 };
-
-export default importBuildTarget;
