+++ src/index.tsx
@@ -3,9 +3,10 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { ErrorBoundary } from './components/pages';
 import store from './store/store';
+import { isDevEnv } from './config';
+import { configureAxios } from './config/configure-axios';
+import { importBuildTarget } from './config/import-build-target';
 import reportWebVitals from './reportWebVitals';
-import { configureAxios } from './api/configureAxios';
-import importBuildTarget from './config/import-build-target';
 import './styles/index.scss';
 
 configureAxios();
@@ -27,4 +28,6 @@ importBuildTarget()
     /* TODO: показать окошко, что все пропало */
   });
 
-reportWebVitals();
+if (isDevEnv) {
+  reportWebVitals();
+}
